+++ src/main/java/com/elksandro/seunegocio/models/item/dto/ItemStockUpdate.java
@@ -6,5 +6,5 @@ import jakarta.validation.constraints.PositiveOrZero;
 public record ItemStockUpdate(
     @NotNull(message = "A quantidade em estoque é obrigatória.")
     @PositiveOrZero(message = "A quantidade deve ser zero ou positiva.")
-    Long stockQuantity
+    Integer stockQuantity
 ) {}
